backend/requirements.txt
flask
requests
flask-cors
mysql-connector-python
pytz